refactor(#8): Use MessageBuilder for message construction
