+++ internal/protocol/jsonrpc_test.go
@@ -11,17 +11,11 @@ import (
 func TestJSONRPCResponse_UnmarshalMessage(t *testing.T) {
 	before := JSONRPCResponse{
 		ID: float64(1),
-		Result: Message{
-			Role: "agent",
-			Parts: []Part{
-				&TextPart{
-					Kind: PartKindText,
-					Text: "Why did the chicken cross the road? To get to the other side!",
-				},
-			},
-			MessageID: "363422be-b0f9-4692-a24d-278670e7c7f1",
-			Kind:      KindMessage,
-		},
+		Result: NewMessageBuilder().
+			MessageID("363422be-b0f9-4692-a24d-278670e7c7f1").
+			Role("agent").
+			Part(NewText("Why did the chicken cross the road? To get to the other side!")).
+			Build(),
 	}
 	var after JSONRPCResponse
